pyproject.toml
[tool.poetry]
name = "testrepl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
Flask-Session = "^0.4.0"
replit = "^3.2.4"
psycopg-binary = "^3.0.1"
psycopg2-binary = "^2.9.1"
implicit = "^0.4.8"
ml_metrics = "^0.1.4"
pandas = "^1.3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
